fix: reestruturação das telas em componentes
