+++ app/(auth)/login/index.tsx
@@ -1,0 +1,5 @@
+import Login from "@/components/login_page/login";
+
+export default function rotaLogin(){
+    return <Login />;
+}
