+++ app/(main)/(tabs)/automacao/index.tsx
