+++ app/(main)/perfil/index.tsx
@@ -1,0 +1,5 @@
+import Perfil from "@/components/perfil_page/perfil";
+
+export default function rotaPerfil(){
+    return <Perfil />;
+}
